Bids and services were moved into sub-directories
