+++ public/mix-manifest.json
@@ -1,5 +1,5 @@
 {
-  "/js/app.js": "/js/app.e948e28cad637c9389b5.js",
+  "/js/app.js": "/js/app.ed60e65aa736a3385445.js",
   "/css/app.css": "/css/app.d5663409871baf9e01819490ba9cef21.css",
   "/js/vendor.js": "/js/vendor.896a8af1e3fc2f4ac84f.js",
   "/js/manifest.js": "/js/manifest.d41d8cd98f00b204e980.js"
